modify port from uint64 to uint16

+++ cluster/rpc/grpc_test.go
@@ -9,10 +9,10 @@ import (
 	"github.com/ethereum/go-ethereum/rpc"
 )
 
-func testSlaveConfig(idx int) *config.SlaveConfig {
+func testSlaveConfig(idx uint16) *config.SlaveConfig {
 	return &config.SlaveConfig{
 		IP:            "127.0.0.1",
-		Port:          uint64(config.SlavePort + idx),
+		Port:          config.SlavePort + idx,
 		ID:            "S" + string(idx),
 		ChainMaskList: nil,
 	}
